requirements.txt
streamlit==1.43.2
google-genai==1.5.0